// src/generated/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
// ...
export interface Scalars {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  DateTime: { input: any; output: any; }
  JSONObject: { input: any; output: any; }
  Time: { input: any; output: any; }
}

export interface ApiError extends Error {
  __typename?: 'ApiError';
  args?: Maybe<Scalars['JSONObject']['output']>;
  code?: Maybe<Scalars['Int']['output']>;
  data?: Maybe<Scalars['JSONObject']['output']>;
  message?: Maybe<Scalars['String']['output']>;
  path?: Maybe<Scalars['String']['output']>;
}
// ...
export interface Error {
  code?: Maybe<Scalars['Int']['output']>;
  data?: Maybe<Scalars['JSONObject']['output']>;
  message?: Maybe<Scalars['String']['output']>;
}
// ...
export interface PermissionRequest {
  /**
   * objectName maps to name / path suffix of the template group representing the permission in Zabbix:
   * Permissions/{objectName}
   */
  objectName: Scalars['String']['input'];
  permission: Permission;
}
// ...
export interface QueryHasPermissionsArgs {
  permissions: Array<PermissionRequest>;
}
// ...
export interface UserPermission {
  __typename?: 'UserPermission';
  /**
   * objectName maps to name / path suffix of the template group representing the permission in Zabbix:
   * Permissions/{objectName}
   */
  objectName: Scalars['String']['output'];
  permission: Permission;
}